backend/requirements.txt
# ── GraphQL
strawberry-graphql[fastapi]>=0.240.0

# ── Web server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
gunicorn>=22.0.0

# ── Lambda adapter
mangum>=0.17.0

# ── AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# ── Auth / JWT
python-jose[cryptography]>=3.3.0
httpx>=0.27.0

# ── PDF generation
weasyprint>=61.0
jinja2>=3.1.4

# ── Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.9
